+++ src/FlatSharp.Runtime/IO/SerializationTarget/IFlatBufferSerializationTarget.cs
@@ -14,8 +14,6 @@
  * limitations under the License.
  */
 
-using System.Buffers.Binary;
-
 namespace FlatSharp;
 
 /// <summary>
@@ -51,140 +49,4 @@ public interface IFlatBufferSerializationTarget<out T>
     /// Returns a Span{byte} over the given range.
     /// </summary>
     Span<byte> AsSpan(long start, int length);
-}
-
-/// <summary>
-/// Extensions for IFlatBufferSerializationTarget
-/// </summary>
-public static class FlatBufferSerializationTargetExtensions
-{
-    public static void WriteBool<T>(this T target, long offset, bool value)
-        where T : IFlatBufferSerializationTarget<T>
-#if NET9_0_OR_GREATER
-        , allows ref struct
-#endif
-    {
-        target.WriteUInt8(offset, value ? SerializationHelpers.True : SerializationHelpers.False);
-    }
-    
-    public static void WriteUInt8<T>(this T target, long offset, byte value)
-        where T : IFlatBufferSerializationTarget<T>
-    #if NET9_0_OR_GREATER
-        , allows ref struct
-    #endif
-    {
-        target[offset] = value;
-    }
-    
-    public static void WriteInt8<T>(this T target, long offset, sbyte value)
-        where T : IFlatBufferSerializationTarget<T>
-#if NET9_0_OR_GREATER
-        , allows ref struct
-#endif
-    {
-        target[offset] = unchecked((byte)value);
-    }
-    
-    public static void WriteUInt16<T>(this T target, long offset, ushort value)
-        where T : IFlatBufferSerializationTarget<T>
-#if NET9_0_OR_GREATER
-        , allows ref struct
-#endif
-    {
-        BinaryPrimitives.WriteUInt16LittleEndian(
-            target.AsSpan(offset, sizeof(ushort)),
-            value);
-    }
-    
-    public static void WriteInt16<T>(this T target, long offset, short value)
-        where T : IFlatBufferSerializationTarget<T>
-#if NET9_0_OR_GREATER
-        , allows ref struct
-#endif
-    {
-        BinaryPrimitives.WriteInt16LittleEndian(
-            target.AsSpan(offset, sizeof(short)),
-            value);
-    }
-    
-    public static void WriteUInt32<T>(this T target, long offset, uint value)
-        where T : IFlatBufferSerializationTarget<T>
-#if NET9_0_OR_GREATER
-        , allows ref struct
-#endif
-    {
-        BinaryPrimitives.WriteUInt32LittleEndian(
-            target.AsSpan(offset, sizeof(uint)),
-            value);
-    }
-    
-    public static void WriteInt32<T>(this T target, long offset, int value)
-        where T : IFlatBufferSerializationTarget<T>
-#if NET9_0_OR_GREATER
-        , allows ref struct
-#endif
-    {
-        BinaryPrimitives.WriteInt32LittleEndian(
-            target.AsSpan(offset, sizeof(int)),
-            value);
-    }
-    
-    public static void WriteUInt64<T>(this T target, long offset, ulong value)
-        where T : IFlatBufferSerializationTarget<T>
-#if NET9_0_OR_GREATER
-        , allows ref struct
-#endif
-    {
-        BinaryPrimitives.WriteUInt64LittleEndian(
-            target.AsSpan(offset, sizeof(ulong)),
-            value);
-    }
-    
-    public static void WriteInt64<T>(this T target, long offset, long value)
-        where T : IFlatBufferSerializationTarget<T>
-#if NET9_0_OR_GREATER
-        , allows ref struct
-#endif
-    {
-        BinaryPrimitives.WriteInt64LittleEndian(
-            target.AsSpan(offset, sizeof(long)),
-            value);
-    }
-    
-    
-    public static void WriteFloat32<T>(this T target, long offset, float value)
-        where T : IFlatBufferSerializationTarget<T>
-#if NET9_0_OR_GREATER
-        , allows ref struct
-#endif
-    {
-        
-#if NETSTANDARD
-        ScalarSpanReader.FloatLayout floatLayout = new ScalarSpanReader.FloatLayout
-        {
-            value = value
-        };
-        
-        target.WriteUInt32(offset, floatLayout.bytes);
-#else
-        BinaryPrimitives.WriteSingleLittleEndian(
-            target.AsSpan(offset, sizeof(float)),
-            value);
-#endif
-    }
-    
-    public static void WriteFloat64<T>(this T target, long offset, double value)
-        where T : IFlatBufferSerializationTarget<T>
-#if NET9_0_OR_GREATER
-        , allows ref struct
-#endif
-    {
-#if NETSTANDARD
-        target.WriteInt64(offset, BitConverter.DoubleToInt64Bits(value));
-#else
-        BinaryPrimitives.WriteDoubleLittleEndian(
-            target.AsSpan(offset, sizeof(double)),
-            value);  
-#endif   
-    }
 }
