+++ src/FlatSharp/TypeModel/Vectors/MemoryVectorTypeModel.cs
@@ -78,7 +78,7 @@ public class MemoryVectorTypeModel : BaseVectorTypeModel
     public override CodeGeneratedMethod CreateSerializeMethodBody(SerializationCodeGenContext context)
     {
         string body =
-            $"{context.TargetVariableName}.{nameof(SpanWriterExtensions.WriteReadOnlyByteMemoryBlock)}({context.ValueVariableName}, {context.OffsetVariableName}, {context.SerializationContextVariableName});";
+            $"{context.TargetVariableName}.WriteReadOnlyByteMemoryBlock({context.ValueVariableName}, {context.OffsetVariableName}, {context.SerializationContextVariableName});";
 
         return new CodeGeneratedMethod(body);
     }
