+++ src/Tests/FlatSharpEndToEndTests/Oracle/OracleSerializeTests.cs
@@ -45,10 +45,10 @@ public class OracleSerializeTests
             UShort = GetRandom<ushort>(),
         };
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.BasicTypes.Serializer.Write(memory, simple);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.BasicTypesVerify.Verify(new(bb), 4));
         var oracle = Flatc.BasicTypes.GetRootAsBasicTypes(bb);
 
@@ -74,10 +74,10 @@ public class OracleSerializeTests
     {
         var simple = new FS.BasicTypes();
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.BasicTypes.Serializer.Write(memory, simple);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.BasicTypesVerify.Verify(new(bb), 4));
         var oracle = Flatc.BasicTypes.GetRootAsBasicTypes(bb);
 
@@ -111,10 +111,10 @@ public class OracleSerializeTests
             }
         };
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.LinkedListNode.Serializer.Write(memory, simple);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.LinkedListNodeVerify.Verify(new(bb), 4));
         var oracle = Flatc.LinkedListNode.GetRootAsLinkedListNode(bb);
 
@@ -138,10 +138,10 @@ public class OracleSerializeTests
             },
         };
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.LocationHolder.Serializer.Write(memory, holder);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.LocationHolderVerify.Verify(new(bb), 4));
         var oracle = Flatc.LocationHolder.GetRootAsLocationHolder(bb);
 
@@ -176,10 +176,10 @@ public class OracleSerializeTests
             ByteVector3 = new byte[0],
         };
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.Vectors.Serializer.Write(memory, table);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.VectorsVerify.Verify(new(bb), 4));
         var oracle = Flatc.Vectors.GetRootAsVectors(bb);
 
@@ -223,11 +223,11 @@ public class OracleSerializeTests
             }
         };
 
-        Span<byte> buffer = new byte[1024];
+        byte[] buffer = new byte[1024];
         var bytecount = FS.FiveByteStructTable.Serializer.Write(buffer, table);
-        buffer = buffer.Slice(0, (int)bytecount);
+        var span = buffer.AsSpan().Slice(0, (int)bytecount);
 
-        var bb = new ByteBuffer(buffer.ToArray());
+        var bb = new ByteBuffer(span.ToArray());
         Assert.IsTrue(Flatc.FiveByteStructTableVerify.Verify(new Verifier(bb), 4));
         var oracle = Flatc.FiveByteStructTable.GetRootAsFiveByteStructTable(bb);
 
@@ -267,10 +267,10 @@ public class OracleSerializeTests
             Value = new FS.Union(simple)
         };
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.UnionTable.Serializer.Write(memory, table);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.UnionTableVerify.Verify(new(bb), 4));
         var oracle = Flatc.UnionTable.GetRootAsUnionTable(bb);
 
@@ -303,10 +303,10 @@ public class OracleSerializeTests
             Value = new("foobar")
         };
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.UnionTable.Serializer.Write(memory, table);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.UnionTableVerify.Verify(new(bb), 4));
         var oracle = Flatc.UnionTable.GetRootAsUnionTable(bb);
 
@@ -322,10 +322,10 @@ public class OracleSerializeTests
             Value = null,
         };
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.UnionTable.Serializer.Write(memory, table);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.UnionTableVerify.Verify(new(bb), 4));
         var oracle = Flatc.UnionTable.GetRootAsUnionTable(bb);
 
@@ -347,10 +347,10 @@ public class OracleSerializeTests
             Value = new(location)
         };
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.UnionTable.Serializer.Write(memory, table);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.UnionTableVerify.Verify(new(bb), 4));
         var oracle = Flatc.UnionTable.GetRootAsUnionTable(bb);
 
@@ -374,10 +374,10 @@ public class OracleSerializeTests
             }
         };
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.NestedStructs.Serializer.Write(memory, nested);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.NestedStructsVerify.Verify(new(bb), 4));
         var oracle = Flatc.NestedStructs.GetRootAsNestedStructs(bb);
 
@@ -400,10 +400,10 @@ public class OracleSerializeTests
             }
         };
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.NestedStructs.Serializer.Write(memory, nested);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.NestedStructsVerify.Verify(new(bb), 4));
         var oracle = Flatc.NestedStructs.GetRootAsNestedStructs(bb);
 
@@ -427,10 +427,10 @@ public class OracleSerializeTests
             }
         };
 
-        Span<byte> memory = new byte[10240];
+        byte[] memory = new byte[10240];
         long size = FS.VectorOfUnionTable.Serializer.Write(memory, table);
 
-        var bb = new ByteBuffer(memory.Slice(0, (int)size).ToArray());
+        var bb = new ByteBuffer(memory.AsSpan().Slice(0, (int)size).ToArray());
         Assert.IsTrue(Flatc.UnionVectorTableVerify.Verify(new(bb), 4));
         var oracle = Flatc.UnionVectorTable.GetRootAsUnionVectorTable(bb);
